Moved signup sheet into own component
