+++ app/components/form/formControl/Form.tsx
@@ -4,8 +4,8 @@ import { FC, useState } from "react"
 import { useForm, FormProvider } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import * as z from "zod"
-import LoadTemplate from "@template/LoadTemplate"
-import { ProductNameProps } from "@template/templateTypes"
+import LoadTemplate from "@components/templates/LoadTemplate"
+import { ProductNameProps } from "@components/templates/templateTypes"
 import ModalAlert from "@ui/ModalAlert"
 import ModalViewDoc from "@ui/ModalViewDoc"
 import CardSection from "@ui/CardSection"
