+++ app/components/form/formControl/FormTest2.tsx
@@ -4,7 +4,7 @@ import React, { FC, useState } from "react"
 import { useForm, FormProvider } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import * as z from "zod"
-import { ProductNameProps } from "@template/templateTypes"
+import { ProductNameProps } from "@components/templates/templateTypes"
 import ModalAlert from "@ui/ModalAlert"
 import ModalViewDoc from "@ui/ModalViewDoc"
 import CardSection from "@ui/CardSection"
@@ -17,12 +17,7 @@ import PillBar from "@form/PillBar"
 import { Wrench } from "lucide-react"
 import SubmitButton from "@form/SubmitButton"
 import Divider from "@ui/Divider"
-import ModalSheet from "@ui/ModalSheet"
-import Heading from "@ui/Heading"
-import Paragraph from "@ui/Paragraph"
-import Field from "@formControl/Field"
-import Input from "@form/Input"
-import Checkbox from "@form/Checkbox"
+import SheetSignUp from "@uiTemplates/SheetSignUp"
 
 export type FormProps = {
   id: string
@@ -67,58 +62,7 @@ const FormTest2: FC<FormProps> = ({
         {/* <SubmitButton onSubmit={onSubmit} formData={formData}/> */}
 
         {/* --------- Testing sign up modal --------------- */}
-        <div className="flex justify-center gap-[25px] mt-10 mb-28 ">
-          <ModalSheet
-            triggerComponent={
-              <button className="items-center justify-center inline-flex h-10 px-4 font-medium text-slate-500 bg-slate-200 hover:bg-slate-300 border-2 border-slate-400 rounded-lg leading-none outline-none focus:shadow-[0_0_0_2px] shadow-md focus:shadow-sky-400">
-                Test
-              </button>
-            }
-            title="Document Title"
-            description="This is a description"
-          >
-            <form
-              id={id}
-              className="lg:col-span-3"
-              onSubmit={methods.handleSubmit(() => {
-                console.log("Submitted contact details")
-              })} // use RHF's handleSubmit to prevent default form submission behavior
-            >
-              <CardSection id="signupModalContent" variant="blank">
-                <Heading size="h2">Last step ...</Heading>
-                <Paragraph>
-                  Your document contains sensitive information. Protect your
-                  privacy by creating a free account to save, edit, and download
-                  your document.
-                </Paragraph>
-                <Field name="userEmail" validateOnBlur={true}>
-                  <Field.GroupLabel>Email:</Field.GroupLabel>
-                  <Field.Tip>
-                    Unsubscribe anytime. Your email is kept private and is only
-                    used to send you the newsletter.
-                  </Field.Tip>
-                  <Field.Control>
-                    <Input type="text" />
-                  </Field.Control>
-                </Field>
-                <Field
-                  name="checkboxAcceptMarketingEmails"
-                  //validateOnBlur={false}
-                >
-                  <Field.GroupLabel>
-                    Toggle checkbox to change PDF content:
-                  </Field.GroupLabel>
-                  <Field.Control>
-                    <Checkbox>
-                      Send me the newsletter for tips & promos
-                    </Checkbox>
-                  </Field.Control>
-                </Field>
-                <button type="submit">Sign Up Test</button>
-              </CardSection>
-            </form>
-          </ModalSheet>
-        </div>
+        <SheetSignUp id="signupSheetTest" />
         {/* ----------------------------- */}
 
         <PillBar variant="standard" />
