+++ app/components/form/formControl/FormTestTemplate.tsx
@@ -4,8 +4,8 @@ import React, { FC, useState } from "react"
 import { useForm, FormProvider } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import * as z from "zod"
-import LoadTemplate from "@template/LoadTemplate"
-import { ProductNameProps } from "@template/templateTypes"
+import LoadTemplate from "@components/templates/LoadTemplate"
+import { ProductNameProps } from "@components/templates/templateTypes"
 import ModalAlert from "@ui/ModalAlert"
 import ModalStandard from "@ui/ModalStandard"
 import ModalViewDoc from "@ui/ModalViewDoc"
