+++ app/components/templates/TemplateGeneric.tsx
