+++ app/components/templates/context.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react"
-import { PageContextType } from "@template/templateTypes"
+import { PageContextType } from "@components/templates/templateTypes"
 
 // Create context for this page to pass needed values to children
 // Most of template functinality should be handled via RHF's FormContext
