+++ app/product1/page.tsx
@@ -3,7 +3,7 @@
 import * as z from "zod"
 
 import Product1Template from "@product1/product1Template.mdx"
-import { PageContextType } from "@template/templateTypes"
+import { PageContextType } from "@components/templates/templateTypes"
 import Heading from "@ui/Heading"
 import Paragraph from "@ui/Paragraph"
 import Form from "@formControl/Form"
@@ -15,7 +15,7 @@ import { useState, createContext } from "react"
 import RadioGroup from "@components/form/RadioGroup"
 import Select from "@components/form/Select"
 import product1SchemaTest from "@product1/product1SchemaTest.json"
-import { PageContext } from "@template/context"
+import { PageContext } from "@components/templates/context"
 import ModalStandard from "@components/ui/ModalStandard"
 import ModalViewDoc from "@components/ui/ModalViewDoc"
 import { PDFViewer, StyleSheet } from "@react-pdf/renderer"
