+++ app/product1/product1Template.mdx
@@ -1,9 +1,9 @@
 import CardSection from "@ui/CardSection"
-import TemplateTest from "@template/TemplateTest"
+import TemplateTest from "@templates/TemplateTest"
 import product1SchemaTest from "./product1SchemaTest"
-import DynamicTemplateContent from "@template/DynamicTemplateContent"
-import DynamicUserContent from "@template/DynamicUserContent"
-import HideContent from "@template/HideContent"
+import DynamicTemplateContent from "@templates/DynamicTemplateContent"
+import DynamicUserContent from "@templates/DynamicUserContent"
+import HideContent from "@templates/HideContent"
 
 ## Product 1
